pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qaml"
version = "0.0.1"
authors = [
    {name = "Miguel Salinas", email = "miguel@camelqa.com"},
]
description = "Control your devices with natural language"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
	"Appium-Python-Client >= 4.0.0",
	"Pillow >= 10.2",
	"requests >= 2.31",
]

[project.urls]
"Homepage" = "https://github.com/qaml-ai/qaml-python"

[project.scripts]
qaml = "qaml.__main__:main"

[tool.setuptools.package-data]
"src" = ["*.txt", "*.dat"]

